backend/requirements.txt
Flask==2.3.3
flask-pymongo==2.3.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pymongo==4.5.0
dnspython==2.4.2
